refactor(api): move oidc identity providers serializer
into identity-access-management bounded context
