+++ api/lib/application/authentication/oidc/oidc-controller.js
@@ -1,6 +1,6 @@
 import { PIX_ADMIN } from '../../../../src/authorization/domain/constants.js';
+import * as oidcProviderSerializer from '../../../../src/identity-access-management/infrastructure/serializers/jsonapi/oidc-identity-providers.serializer.js';
 import { oidcAuthenticationServiceRegistry, usecases } from '../../../domain/usecases/index.js';
-import * as oidcProviderSerializer from '../../../infrastructure/serializers/jsonapi/oidc-identity-providers-serializer.js';
 import * as oidcSerializer from '../../../infrastructure/serializers/jsonapi/oidc-serializer.js';
 import { BadRequestError, UnauthorizedError } from '../../http-errors.js';
